Add back the 'video_shader' parameter.
The `video_shader` configuration parameter has been removed from RetroArch a while ago,
in favour of per-game/core/folder shader overrides.

However, the built-in shader override mechanism depends on the core name used, while `video_shader`
can be used without any core (name) dependency, which makes it easier to configure with the RetroPie
configuration editor.

The patch will load the shader preset specified in the `video_shader` path if none of the built-in shader
overrides have not been found. The parameter is not saved as part of any override/configuration file when
saving is done from RetroArch's menu, the only mechanism to set it is by directly editing the conf file or via the configuration editor.